requirements.txt
flask
openai
python-dotenv
xhtml2pdf
markdown
fpdf